ENH support tf
* fixed inputs to averaging function on both front ends, changed the appropriate backend functions on both backends

* some fixup of the test file, tests fail by 1

* made behavior of numpy backend aggregate method eqivalent to pytorch backend's aggregate method, added checking of the order_0 coefficent in numpy scattering test

* Revert "made behavior of numpy backend aggregate method eqivalent to pytorch backend's aggregate method, added checking of the order_0 coefficent in numpy scattering test"

This reverts commit f9b3a16f52cd6357ea855b474b3617841efa0b33.

* fixed numpy backend finalize method

* fixed typo

* removed unnecessary indexing of array before numpy compute integral

error is now  0.7823316457158547, down from .94-something

* fixed finialize method, numpy tests pass now

* Fixed pytorch tests crashing

When I changed names I forgot to change the input argument from low_pass to filter, and to grab the Jth filter. This is fixed.

* adding all previously made changes. Other tests pass now, tf fails with .04

* supporting tf and fixing a bug with it

+++ kymatio/scattering3d/filter_bank.py
@@ -67,7 +67,7 @@ def gaussian_filter_bank(M, N, O, J, sigma_0, fourier=True):
             torch array array of size (J+1, M, N, O, 2) containing the (J+1)
             Gaussian filters.
     """
-    gaussians = np.zeros((J + 1, M, N, O))
+    gaussians = np.zeros((J + 1, M, N, O), dtype='complex64')
     for j in range(J + 1):
         sigma = sigma_0 * 2 ** j
         gaussian = gaussian_3d(M, N, O, sigma, fourier=fourier)
@@ -98,7 +98,7 @@ def gaussian_3d(M, N, O, sigma, fourier=True):
     grid = np.fft.ifftshift(
         np.mgrid[-M // 2:-M // 2 + M,
                  -N // 2:-N // 2 + N,
-                 -O // 2:-O // 2 + O].astype('float32'),
+                 -O // 2:-O // 2 + O].astype('complex64'),
         axes=(1,2,3))
     _sigma = sigma
     if fourier:
